add unit test for add_owner get_owner services
