+++ .github/workflows/pr-test.yml
@@ -1,4 +1,4 @@
-name: pr test
+name: pr tests
 
 on: [pull_request]
