+++ Dockerfile
@@ -11,5 +11,5 @@ COPY ./src /src
 
 EXPOSE 8000
 
-CMD ["uvicorn", "app.allocation.entrypoints.fastapi_app:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
+CMD ["uvicorn", "app.entrypoints.fastapi_app:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
