--- src/app/allocation/service_layer/unit_of_work.py
@@ -1,42 +1,0 @@
-from __future__ import annotations
-import abc
-from app.allocation.adapters import repository
-
-class AbstractUnitOfWork(abc.ABC):
-
-    batches: repository.AbstractRepository
-
-    def __enter__(self) -> AbstractUnitOfWork:
-        return self
-
-    def __exit__(self, *args):
-        self.rollback()
-
-    @abc.abstractmethod
-    def commit(self):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def rollback(self):
-        raise NotImplementedError
-
-
-class SqlAlchemyUnitOfWork(AbstractUnitOfWork):
-
-    def __init__(self, session_factory):
-        self.session_factory = session_factory
-
-    def __enter__(self):
-        self.session = self.session_factory()
-        self.batches = repository.SqlAlchemyRepository(self.session)
-        return super().__enter__()
-
-    def __exit__(self, *args):
-        super().__exit__(*args)
-        self.session.close()
-
-    def commit(self):
-        self.session.commit()
-
-    def rollback(self):
-        self.session.rollback()
