--- src/app/test/integration/fake_unit_of_work.py
@@ -1,24 +1,0 @@
-from __future__ import annotations
-from app.allocation.service_layer.unit_of_work import AbstractUnitOfWork
-from app.test.integration import fake_repository
-
-
-class FakeUnitOfWork(AbstractUnitOfWork):
-
-    def __init__(self, session_factory):
-        self.session_factory = session_factory
-
-    def __enter__(self):
-        self.session = self.session_factory()
-        self.batches = fake_repository.FakeRepository(self.session)
-        return super().__enter__()
-
-    def __exit__(self, *args):
-        super().__exit__(*args)
-        self.session.close()
-
-    def commit(self):
-        self.session.commit()
-
-    def rollback(self):
-        self.session.rollback()
