--- src/app/test/unit/owner/conftest.py
@@ -1,4 +1,0 @@
-from app.allocation.domain.schemas import Owner
-
-
-
